third_party/blink/web_tests/virtual/origin-trials-runtimeflags-disabled/http/tests/origin_trials/webexposed/fetch-later-origin-trail-interfaces-expected.txt

This is a testharness.js-based test.
FAIL FetchLater API interface in Origin-Trial enabled document. assert_equals: Interface fetchLater exists on provided object expected true but got false
Harness: the test ran to completion.
